requirements.txt
apache-airflow-providers-docker
apache-airflow-providers-google